eventmap-fastapi/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openrouteservice
qdrant-client
geopandas
pandas
shapely
folium
jinja2
python-multipart